Moved add balance functionality to load_balances
